refactor(marketAdmin): extract types and mock data to use production API
